// src/app/models/etudiant.ts
export class Apprenants {
  idApprenant: number = 0;
  nomApprenant: string = '';
  prenomApprenant: string = '';
  emailApprenant: string = '';
  passwordApprenant: string = '';
  imageApprenants: string = '';
  etatApprenant: number = 0;
  role: string = '';
  anneeScolaire: Date = new Date();
  niveau: string = '';
}